requirements.txt
torch>=1.8.0
torchvision>=0.9.0
torchaudio>=0.8.0

megengine-cvpods @ git+https://github.com/Megvii-BaseDetection/cvpods.git
fvcore>=0.1.5
iopath>=0.1.7
omegaconf>=2.1.0
hydra-core>=1.1.0
cloudpickle>=1.6.0
tabulate>=0.8.9

opencv-python>=4.5.0
Pillow>=8.0.0
albumentations>=1.0.0
pycocotools>=2.0.2

numpy>=1.19.0
scipy>=1.6.0
matplotlib>=3.3.0
seaborn>=0.11.0
pandas>=1.2.0
tqdm>=4.60.0
pyyaml>=5.4.0
termcolor>=1.1.0

plotly>=5.0.0
wandb>=0.12.0

psutil>=5.8.0
GPUtil>=1.4.0